added static food shape
This is weird. I added in a food shape, a diagonal path from bottom
left to top right. Now all the agents are following it. Even the ones
which were trained on a diagonal the inverse of it. It probably means
I’ve not done the JSON switch statement correctly.

What it ended up being was, the switch statement wasn’t working and
even if it had been the JSON for that agent is corrupted somehow.