new way of training the agents
Since we no longer care about the snake colliding with itself, purely about patterns, I thought I might try out the suggestion made by Andrej Karpathy http://cs.stanford.edu/people/karpathy/convnetjs/?place=msg%2Fconvnetjs%2FC3JggTC2pHE%2FfkMV7_CJCwAJ

This entails having separate matrices for each part of the snake and a separate one for the food. If we were to this properly even for the 5x5 matrix we would need 5*5*26 rows in an input vector, which is just way too many. However if we limit ourselves to just the head of the snake and the item of food, we get a very manageable 5*5*2. Trying this out with 3 layers of 100 nodes initially to see how it gets on.