fixed backwards reward function
Now the agent gets rewarded after eating the food, rather than for the
action preceding it. By giving the agent the reward before stepping on
the food, it causing agents to learn to avoid the food but travel
close. The new agents create quite nice and complex patterns, even with
with very sparse network architecture.