An extremely well trained agent
Fair enough, it only works when the agent can't eat itself but after 33 million iterations it's extremely efficient at eating food items. This is even with giving reward for the action of moving onto the food occupied space. It just took a lot of training to work. We should put back the traditional SARSA in the newer version of the separate encoding training file. Which as usual, I've forgotten to hit sync with the local github app.